refactor styler tests
